+++ ext/js/core/log.js
