+++ ext/js/dictionary/dictionary-database.js
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import {log} from '../core/logger.js';
+import {log} from '../core/log.js';
 import {stringReverse} from '../core/utilities.js';
 import {Database} from '../data/database.js';
