+++ ext/js/dictionary/dictionary-worker-main.js
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import {log} from '../core/logger.js';
+import {log} from '../core/log.js';
 import {DictionaryWorkerHandler} from './dictionary-worker-handler.js';
 
 /** Entry point. */
